bytearray, codecs: follow the slot storage and encode error signatures
`slots::slot_get`/`slot_del` now take the slot storage value instead of
`(obj, index, slots_field)`, and `slot_set` is the `slot_set_direct!`
expansion, so the bytearray helpers read `w_slots` off the layout the way
the float and complex helpers do.  `unicode_encode_error` takes `i64`
bounds.

+++ pyre/pyre-interpreter/src/module/_multibytecodec/mod.rs
@@ -213,8 +213,8 @@ fn encode_rust_codec_with_state(
                         return Err(crate::typedef::unicode_encode_error(
                             name,
                             roots.get(input_slot),
-                            position,
-                            end,
+                            position as i64,
+                            end as i64,
                             "illegal multibyte sequence",
                         ));
                     }
